+++ src/gui/gui2_textentry.cpp
@@ -184,7 +184,7 @@ void GuiTextEntry::onTextInput(const string& text)
     }
 
     this->text = this->text.substr(0, std::min(selection_start, selection_end)) + text + this->text.substr(std::max(selection_start, selection_end));
-    selection_end = selection_start = std::min(selection_start, selection_end) + text.length();
+    selection_end = selection_start = static_cast<int>(std::min(selection_start, selection_end) + text.length());
     runChangeCallback();
 }
 
@@ -263,7 +263,7 @@ void GuiTextEntry::onTextInput(sp::TextInputEvent e)
             if (end_of_current_line < 0) return;
 
             int end_of_end_line = text.find("\n", end_of_current_line + 1);
-            if (end_of_end_line == -1) end_of_end_line = text.length();
+            if (end_of_end_line == -1) end_of_end_line = static_cast<int>(text.length());
 
             const int offset = selection_end - start_of_current_line;
             selection_end = end_of_current_line + 1 + std::min(offset, end_of_end_line - (end_of_current_line + 1));
@@ -281,7 +281,7 @@ void GuiTextEntry::onTextInput(sp::TextInputEvent e)
     case sp::TextInputEvent::LineEnd:
     case sp::TextInputEvent::LineEndWithSelection:
         selection_end = text.find("\n", selection_start);
-        if (selection_end == -1) selection_end = text.length();
+        if (selection_end == -1) selection_end = static_cast<int>(text.length());
         if (e != sp::TextInputEvent::LineEndWithSelection)
             selection_start = selection_end;
         break;
@@ -293,13 +293,13 @@ void GuiTextEntry::onTextInput(sp::TextInputEvent e)
         break;
     case sp::TextInputEvent::TextEnd:
     case sp::TextInputEvent::TextEndWithSelection:
-        selection_end = text.length();
+        selection_end = static_cast<int>(text.length());
         if (e != sp::TextInputEvent::TextEndWithSelection)
             selection_start = selection_end;
         break;
     case sp::TextInputEvent::SelectAll:
         selection_end = 0;
-        selection_start = text.length();
+        selection_start = static_cast<int>(text.length());
         break;
     case sp::TextInputEvent::Delete:
         if (readonly) return;
@@ -343,7 +343,7 @@ void GuiTextEntry::onTextInput(sp::TextInputEvent e)
             int start_of_line = text.substr(0, std::min(selection_start, selection_end)).rfind("\n") + 1;
             auto data = text.substr(start_of_line, std::max(selection_start, selection_end));
             data = "    " + data.replace("\n", "\n    ");
-            int extra_length = data.length() - (std::max(selection_start, selection_end) - start_of_line) - 4;
+            int extra_length = static_cast<int>(data.length()) - (std::max(selection_start, selection_end) - start_of_line) - 4;
             text = text.substr(0, start_of_line) + data + text.substr(std::max(selection_start, selection_end));
 
             if (start_of_line != selection_start)
@@ -371,7 +371,7 @@ void GuiTextEntry::onTextInput(sp::TextInputEvent e)
                 data = data.replace("\n ", "\n");
             }
 
-            const int removed_length = (std::max(selection_start, selection_end) - start_of_line) - data.length();
+            const int removed_length = static_cast<int>((std::max(selection_start, selection_end) - start_of_line) - data.length());
             text = text.substr(0, start_of_line) + data + text.substr(std::max(selection_start, selection_end));
 
             if (selection_start > selection_end)
@@ -416,7 +416,7 @@ void GuiTextEntry::onFocusGained()
     if (select_on_focus)
     {
         selection_end = 0;
-        selection_start = text.length();
+        selection_start = static_cast<int>(text.length());
     }
 
     typing_indicator = true;
@@ -526,7 +526,7 @@ int GuiTextEntry::getTextOffsetForPosition(glm::vec2 position)
 {
     position = position - rect.position - render_offset;
     position.x -= 16.0f;
-    int result = text.size();
+    int result = static_cast<int>(text.size());
 
     const auto& front = front_style->get(getState());
     // if (vertical_scroll) position.y -= vertical_scroll->getValue();
@@ -556,7 +556,7 @@ int GuiTextEntry::getTextOffsetForPosition(glm::vec2 position)
     }
 
     if (n == pfs.data.size())
-        return text.size();
+        return static_cast<int>(text.size());
 
     float line_y = pfs.data[n].position.y;
     for (; n < pfs.data.size(); n++)
