+++ src/script/components.cpp
@@ -102,12 +102,12 @@
         } \
     };
 #define BIND_ARRAY(T, A) \
-    sp::script::ComponentHandler<T>::array_count_func = [](const T& t) -> int { return t.A.size(); }; \
+    sp::script::ComponentHandler<T>::array_count_func = [](const T& t) -> int { return static_cast<int>(t.A.size()); }; \
     sp::script::ComponentHandler<T>::array_resize_func = [](T& t, int new_size) { t.A.resize(new_size); }; \
     sp::script::ComponentHandler<T>::indexed_members["length"] = { \
         [](lua_State* L, const void* ptr, int n) { \
             auto t = reinterpret_cast<const T*>(ptr); \
-            return sp::script::Convert<int>::toLua(L, t->A.size()); \
+            return sp::script::Convert<int>::toLua(L, static_cast<int>(t->A.size())); \
         }, [](lua_State* L, void* ptr, int n) { \
             auto t = reinterpret_cast<T*>(ptr); \
             t->A.resize(std::max(0, sp::script::Convert<int>::fromLua(L, -1))); \
@@ -146,7 +146,7 @@
         } \
     };
 #define BIND_ARRAY_DIRTY_FLAG(T, A, DIRTY) \
-    sp::script::ComponentHandler<T>::array_count_func = [](const T& t) -> int { return t.A.size(); }; \
+    sp::script::ComponentHandler<T>::array_count_func = [](const T& t) -> int { return static_cast<int>(t.A.size()); }; \
     sp::script::ComponentHandler<T>::array_resize_func = [](T& t, int new_size) { t.A.resize(new_size); t.DIRTY = true; };
 #define BIND_ARRAY_DIRTY_FLAG_MEMBER(T, A, MEMBER, DIRTY) \
     sp::script::ComponentHandler<T>::indexed_members[STRINGIFY(MEMBER)] = { \
@@ -468,7 +468,7 @@ void initComponentScriptBindings()
     sp::script::ComponentHandler<DockingBay>::members["external_dock_classes"] = {
         [](lua_State* L, const void* ptr) {
             auto bay = reinterpret_cast<const DockingBay*>(ptr);
-            lua_createtable(L, bay->external_dock_classes.size(), 0);
+            lua_createtable(L, static_cast<int>(bay->external_dock_classes.size()), 0);
             int idx = 1;
             for(const auto& c : bay->external_dock_classes) {
                 lua_pushstring(L, c.c_str());
@@ -491,7 +491,7 @@ void initComponentScriptBindings()
     sp::script::ComponentHandler<DockingBay>::members["internal_dock_classes"] = {
         [](lua_State* L, const void* ptr) {
             auto bay = reinterpret_cast<const DockingBay*>(ptr);
-            lua_createtable(L, bay->internal_dock_classes.size(), 0);
+            lua_createtable(L, static_cast<int>(bay->internal_dock_classes.size()), 0);
             int idx = 1;
             for(const auto& c : bay->internal_dock_classes) {
                 lua_pushstring(L, c.c_str());
@@ -514,7 +514,7 @@ void initComponentScriptBindings()
     sp::script::ComponentHandler<DockingBay>::members["berths"] = {
         [](lua_State* L, const void* ptr) {
             auto bay = reinterpret_cast<const DockingBay*>(ptr);
-            lua_createtable(L, bay->berths.size(), 0);
+            lua_createtable(L, static_cast<int>(bay->berths.size()), 0);
             int idx = 1;
             for(const auto& berth : bay->berths) {
                 lua_createtable(L, 0, 6);
@@ -560,17 +560,17 @@ void initComponentScriptBindings()
 
                         lua_getfield(L, -1, "move_time");
                         if (lua_isnumber(L, -1))
-                            berth.move_time = lua_tonumber(L, -1);
+                            berth.move_time = static_cast<float>(lua_tonumber(L, -1));
                         lua_pop(L, 1);
 
                         lua_getfield(L, -1, "move_progress");
                         if (lua_isnumber(L, -1))
-                            berth.move_progress = lua_tonumber(L, -1);
+                            berth.move_progress = static_cast<float>(lua_tonumber(L, -1));
                         lua_pop(L, 1);
 
                         lua_getfield(L, -1, "transfer_rate");
                         if (lua_isnumber(L, -1))
-                            berth.transfer_rate = lua_tonumber(L, -1);
+                            berth.transfer_rate = static_cast<float>(lua_tonumber(L, -1));
                         lua_pop(L, 1);
 
                         lua_getfield(L, -1, "transfer_direction");
